+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added support for Stratum 0 and Stratum 1 on Ubuntu hosts.
 - Created setup-example.sh script to simplify testing.
+- Added setup-example.sh and "setup-all" command to automate testing.
 
 ## 1.0.0